+++ tests/unit/tasks/test_structure_run_task.py
@@ -5,7 +5,7 @@ from tests.mocks.mock_prompt_driver import MockPromptDriver
 
 
 class TestStructureRunTask:
-    def test_run(self, mock_config):
+    def test_run_single_input(self, mock_config):
         mock_config.drivers_config.prompt_driver = MockPromptDriver(mock_output="agent mock output")
         agent = Agent()
         mock_config.drivers_config.prompt_driver = MockPromptDriver(mock_output="pipeline mock output")
@@ -13,6 +13,19 @@ class TestStructureRunTask:
         driver = LocalStructureRunDriver(structure_factory_fn=lambda: agent)
 
         task = StructureRunTask(driver=driver)
+
+        pipeline.add_task(task)
+
+        assert task.run().to_text() == "agent mock output"
+
+    def test_run_multiple_inputs(self, mock_config):
+        mock_config.drivers_config.prompt_driver = MockPromptDriver(mock_output="agent mock output")
+        agent = Agent()
+        mock_config.drivers_config.prompt_driver = MockPromptDriver(mock_output="pipeline mock output")
+        pipeline = Pipeline()
+        driver = LocalStructureRunDriver(structure_factory_fn=lambda: agent)
+
+        task = StructureRunTask(input=["foo", "bar", "baz"], driver=driver)
 
         pipeline.add_task(task)